change header prototype
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@19740 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 0d64f53f32c8d4bba6b37f9f369cf841d1b32729

+++ sandbox/point_cloud_clustering/src/generic_clustering.cpp
@@ -70,7 +70,7 @@ int point_cloud_clustering::GenericClustering::computeClusterCentroids(const rob
     }
 
     // normalize by number of points
-    curr_nbr_pts = std::max(curr_nbr_pts, 1);
+    curr_nbr_pts = std::max(curr_nbr_pts, static_cast<unsigned int>(1));
     for (unsigned int i = 0 ; i < 3 ; i++)
     {
       curr_centroid[i] /= static_cast<float> (curr_nbr_pts);
